delete deck security, context notes

+++ imports/api/methods/deckMethods.js
@@ -9,7 +9,15 @@ import { deleteDecksCards } from '/imports/api/methods/cardMethods.js'
 Meteor.methods({
 	deleteDeck(deckId) {
 		check(deckId, String)
+
 		if (!this.userId) throw new Meteor.Error('Not authorized.')
+
+		const isUsersDeck = !!DeckCollection.findOne(
+			{ _id: deckId, userId: this.userId },
+			{ fields: { _id: 1 } }
+		)
+		if (!isUsersDeck) throw new Meteor.Error('Access denied.')
+
 		DeckCollection.update({ _id: deckId }, { $set: { deleted: true } })
 		deleteDecksCards(deckId)
 	},
